기존 프론트엔드 본인 인증 코드 수정 백엔드 본인 인증 기능 추가

+++ src/main/java/egovframework/example/sample/web/RoomSetting.java
@@ -2,8 +2,8 @@ package egovframework.example.sample.web;
 
 public class RoomSetting {
 
-    public int ante;
-    public int max;
+    public long ante;
+    public long max;
     public int maxplayer;
     public String useitem;
